Дробит файл main.js на модули
