+++ index.html
@@ -176,7 +176,7 @@
       </div>
     </section>
   </template>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
