+++ js/check-length-string.js
@@ -1,0 +1,6 @@
+function isAllowedLength(string, maxValue) {
+  return String(string).length <= maxValue;
+}
+isAllowedLength('keks', 9);
+
+export { isAllowedLength };
