+++ js/util.js
@@ -1,0 +1,23 @@
+// isNaN() - функция, определяющая не число. Отдает булевый признак
+function getRandomPositiveInteger(a, b) {
+
+  if (a < 0 || b < 0) {
+    return NaN;
+  }
+
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+}
+
+function getRandomElement(elements) {
+  if (elements.length === 0) {
+    return 'Длина массива = 0';
+  }
+  return elements[getRandomPositiveInteger(0, elements.length - 1)];
+}
+
+export { getRandomElement, getRandomPositiveInteger };
